feat: implemented module task
